Patch reveal and claim actions
