+++ app/actions/chompResult.ts
@@ -1,9 +1,11 @@
 "use server";
 
-import { ResultType } from "@prisma/client";
+import { FungibleAsset, ResultType } from "@prisma/client";
 import { revalidatePath } from "next/cache";
 import prisma from "../services/prisma";
+import { calculateRevealPoints } from "../utils/points";
 import { getQuestionState, isEntityRevealable } from "../utils/question";
+import { incrementFungibleAssetBalance } from "./fungible-asset";
 import { getJwtPayload } from "./jwt";
 
 export async function revealDeck(deckId: number) {
@@ -113,7 +115,7 @@ export async function revealQuestions(questionIds: number[]) {
       .filter((qs) => qs.state.isRevealed)
       .map((qs) => qs.questionId);
 
-    const newelyRevealed = questionStates
+    const newlyRevealed = questionStates
       .filter(
         (qs) =>
           revealableQuestions.some((rq) => rq.id === qs.questionId) &&
@@ -121,14 +123,14 @@ export async function revealQuestions(questionIds: number[]) {
       )
       .map((qs) => qs.questionId);
 
-    const revealedQuestions = [...alreadyRevealed, ...newelyRevealed];
+    const revealedQuestions = [...alreadyRevealed, ...newlyRevealed];
     const allQuestionIds = deck.deckQuestions.map((dq) => dq.questionId);
 
-    const remainigQuestions = allQuestionIds.filter(
+    const remainingQuestions = allQuestionIds.filter(
       (qId) => !revealedQuestions.includes(qId),
     );
 
-    return remainigQuestions.length === 0;
+    return remainingQuestions.length === 0;
   });
 
   await prisma.$transaction(async (tx) => {
@@ -137,15 +139,31 @@ export async function revealQuestions(questionIds: number[]) {
         ...questionIds.map((questionId) => ({
           questionId,
           userId: payload.sub,
-          result: ResultType.Claimed,
+          result: ResultType.Revealed,
         })),
         ...decksToAddRevealFor.map((deck) => ({
           deckId: deck.id,
           userId: payload.sub,
-          result: ResultType.Claimed,
+          result: ResultType.Revealed,
         })),
       ],
     });
+
+    const revealResult = await calculateRevealPoints(
+      payload.sub,
+      questionIds.filter((questionId) => questionId !== null) as number[],
+    );
+
+    const fungibleAssetRevealTasks = revealResult.map((rr) =>
+      incrementFungibleAssetBalance(
+        FungibleAsset.Point,
+        rr.amount,
+        rr.type,
+        tx,
+      ),
+    );
+
+    await Promise.all(fungibleAssetRevealTasks);
   });
 
   revalidatePath("/application");
